+++ app/services/auth/__init__.py
@@ -1,5 +1,5 @@
 """Entry point for the auth service"""
-from .dtos import Credentials, JobStatus
+from .dtos import Credentials, AuthLog, PartialAuthLog
 from .exc import AuthenticationError, AuthorizationError, JobAlreadyExists
 from .service import authenticate, save_credentials
 
@@ -7,7 +7,8 @@ __all__ = [
     save_credentials,
     authenticate,
     Credentials,
-    JobStatus,
+    AuthLog,
+    PartialAuthLog,
     JobAlreadyExists,
     AuthorizationError,
     AuthenticationError,